+++ packaging/linux/app-image/movie-monad-install-app-image-script.sh
@@ -3,12 +3,12 @@
 # (C) 2017 David Lettier
 # lettier.com
 
-_MOVIE_MONAD_VERSION="0.0.6.0"
+_MOVIE_MONAD_VERSION="0.0.6.1"
 _MOVIE_MONAD_RELEASES_DOWNLOAD="https://github.com/lettier/movie-monad/releases/download/$_MOVIE_MONAD_VERSION"
 _MOVIE_MONAD_PACKAGING_LINUX_COMMON="https://raw.githubusercontent.com/lettier/movie-monad/master/packaging/linux/common"
 _MOVIE_MONAD_APP_IMAGE="movie-monad-$_MOVIE_MONAD_VERSION-x86_64.AppImage"
-_MOVIE_MONAD_DESKTOP="movie-monad.desktop"
-_MOVIE_MONAD_ICON="movie-monad-icon.svg"
+_MOVIE_MONAD_DESKTOP="com.lettier.movie-monad.desktop"
+_MOVIE_MONAD_ICON="com.lettier.movie-monad.svg"
 
 echo -e "Installing Movie Monad...\n"